Extract `fetch_*` functions to separate files.
